add basic search framework

+++ app/src/main/java/com/wulee/notebook/ui/MainActivity.java
@@ -31,6 +31,7 @@ import com.wulee.notebook.view.SpacesItemDecoration;
 import com.yanzhenjie.permission.AndPermission;
 import com.yanzhenjie.permission.PermissionListener;
 
+import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 
@@ -65,6 +66,8 @@ public class MainActivity extends BaseActivity implements View.OnClickListener{
     private NoteListAdapter mNoteListAdapter;
     private List<Note> noteList;
     private NoteDao noteDao;
+    private String searchKeyword;
+    private String searchDate;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -90,6 +93,15 @@ public class MainActivity extends BaseActivity implements View.OnClickListener{
         LinearLayoutManager layoutManager = new LinearLayoutManager(this);
         layoutManager.setOrientation(LinearLayoutManager.VERTICAL);//竖向列表
         rv_list_main.setLayoutManager(layoutManager);
+
+        Intent intent = getIntent();
+        try {
+            Bundle bundle = intent.getBundleExtra("data");
+            searchKeyword = (String) bundle.getSerializable("keyword");
+            searchDate = (String) bundle.getSerializable("date");
+        } catch (Exception ex) {
+            // pass
+        }
 
 
         mNoteListAdapter = new NoteListAdapter(R.layout.list_item_note,noteList);
@@ -228,11 +240,9 @@ public class MainActivity extends BaseActivity implements View.OnClickListener{
 
     //刷新笔记列表
     private void refreshNoteList(final boolean isRefresh){
-        UserInfo user = BmobUser.getCurrentUser(UserInfo.class);
-        BmobQuery<Note> query = new BmobQuery<>();
-        query.addWhereEqualTo("user",user);
-        query.include("user");
-        query.order("-createdAt");
+
+        BmobQuery<Note> query = generateQuery();
+
         if(!isRefresh){
             showProgressBar("正在加载...");
         }
@@ -245,22 +255,70 @@ public class MainActivity extends BaseActivity implements View.OnClickListener{
                     swipeLayout.setRefreshing(false);
                 }
                 if(e == null){
-                    noteDao.deleteAllNote();
-                    if(list != null && list.size()>0){
-                        tvNodata.setVisibility(View.GONE);
-                        for(Note note : list){
-                            noteDao.insertNote(note);
+                    if (isSearch()) {
+                        noteDao.deleteAllNote();
+                        if (list != null && list.size() > 0) {
+                            tvNodata.setVisibility(View.GONE);
+                            for (Note note : list) {
+                                noteDao.insertNote(note);
+                            }
                         }
-                    }else{
-                        tvNodata.setVisibility(View.VISIBLE);
                     }
-                    noteList = noteDao.queryNotesAll();
-                    mNoteListAdapter.setNewData(noteList);
+                    if (list == null || list.size() <= 0){
+                            tvNodata.setVisibility(View.VISIBLE);
+                    }
+
+                    //noteList = noteDao.queryNotesAll();
+                    mNoteListAdapter.setNewData(list);
+                    clearSearch();
                 }else{
                     showToast(e.getErrorCode()+ ","+e.getMessage());
                 }
             }
         });
+    }
+
+    /*
+        generate bmob query is search is enabled: fuck! bmob does not provide unpaid user with addWhereContains !!!
+     */
+    private BmobQuery<Note> generateQuery() {
+        UserInfo user = BmobUser.getCurrentUser(UserInfo.class);
+        BmobQuery<Note> query = new BmobQuery<>();
+        query.addWhereEqualTo("user", user);
+
+
+        if (isSearch()) {
+            BmobQuery<Note> queryDate = new BmobQuery<>();
+            BmobQuery<Note> queryTitle = new BmobQuery<>();
+            queryTitle.addWhereEqualTo("user", user);
+            BmobQuery<Note> queryContent = new BmobQuery<>();
+            queryContent.addWhereEqualTo("user", user);
+
+            if (searchKeyword.length() > 0) {
+                //queryTitle.addWhereContains("title", searchKeyword);
+                //queryContent.addWhereContains("content", searchKeyword);
+                //queryTitle = queryTitle.or(Arrays.asList(queryTitle, queryContent));
+                //query = query.and(Arrays.asList(query, queryTitle));
+                query.addWhereEqualTo("title", searchKeyword);
+            }
+            if (searchDate.length() > 0) {
+                query.addWhereEqualTo("updatedAt", searchDate);
+                //query = query.and(Arrays.asList(query, queryDate));
+            }
+
+        }
+        query.include("user");
+        query.order("-createdAt");
+        return query;
+    }
+
+    private void clearSearch() {
+        searchKeyword = null;
+        searchDate = null;
+    }
+
+    private boolean isSearch() {
+        return  (searchKeyword != null || searchDate != null);
     }
 
     @Override
@@ -287,6 +345,11 @@ public class MainActivity extends BaseActivity implements View.OnClickListener{
                 Intent intent = new Intent(MainActivity.this, NewActivity.class);
                 intent.putExtra("flag", 0);
                 startActivity(intent);
+                break;
+            case R.id.action_search:
+                Intent intentSearch = new Intent(MainActivity.this, SearchActivity.class);
+                intentSearch.putExtra("flag", 0);
+                startActivity(intentSearch);
                 break;
         }
         return super.onOptionsItemSelected(item);
